# 8356869: RISC-V: Improve tail handling of array fill stub