@Anjian-Wen 2025-05-13
/issue JDK-8356869