@Hamlin-Li 2025-05-13
Nice catch and fix. Looks good.
Can you add the test case `@param("15")` and maybe some more to ArrayFill.java?