@Anjian-Wen 2025-05-14
@Hamlin-Li Thanks for the review! I have added two sizes to @Param. Size = 15 for this case and size = 7 for the case in JDK-8356593.